Validation

+++ app/Http/Controllers/Admin/ProductController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\ProductRequest;
 use App\Models\Category;
 use App\Models\Product;
 use Illuminate\Http\Request;
@@ -18,6 +19,7 @@ class ProductController extends Controller
     public function index()
     {
         $products = Product::get();
+
         return view('auth.products.index', compact('products'));
     }
 
@@ -28,7 +30,8 @@ class ProductController extends Controller
      */
     public function create()
     {
-        return view('auth.products.form');
+        $categories = Category::get();
+        return view('auth.products.form', compact('categories'));
     }
 
     /**
@@ -37,11 +40,15 @@ class ProductController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(ProductRequest $request)
     {
-        $path=$request->file('image')->store('products');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request->has('image')){
+            $path=$request->file('image')->store('products');
+            $params['image'] = $path;
+        }
+
         Product::create($params);
         return redirect()->route('products.index');
     }
@@ -76,13 +83,15 @@ class ProductController extends Controller
      * @param  \App\Models\Product  $product
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Product $product)
+    public function update(ProductRequest $request, Product $product)
     {
-        Storage::delete($product->image);
-        $path=$request->file('image')->store('products');
         $params = $request->all();
-        $params['image'] = $path;
-
+        unset($params['image']);
+        if ($request->has('image')){
+            Storage::delete($product->image);
+            $path=$request->file('image')->store('products');
+            $params['image'] = $path;
+        }
         $product->update($params);
         return redirect()->route('products.index');
     }
